requirements.txt
flask
torch
torchvision
torchaudio
sentence-transformers
transformers
accelerate
openai-whisper
openai
